Revert unwanted docs/requirementstxt file and Added mean_absolute_error function

+++ ivy/functional/frontends/tensorflow/keras/metrics.py
@@ -216,7 +216,7 @@ logcosh = log_cosh
 
 @to_ivy_arrays_and_back
 def mean_absolute_error(y_true, y_pred):
-    return ivy.mean(ivy.abs(y_true - y_pred), axis=-1).astype(y_true.dtype)
+    return ivy.mean(ivy.abs(y_true - y_pred), axis=-1)
 
 
 mae = mean_absolute_error
